gh-93649: Split time tests from _testcapimodule.c
